requirements.txt
python-dotenv
requests
waitress
Flask
flask-cors
PyJWT
bcrypt
